Revert incorrect parameter argument cast in `loophole_path`

+++ src/xrGame/stalker_movement_manager_smart_cover.cpp
@@ -371,7 +371,7 @@ void stalker_movement_manager_smart_cover::loophole_path(smart_cover::cover cons
 	shared_str target = smart_cover::transform_vertex(target_raw, false);
 
 	typedef GraphEngineSpace::CBaseParameters CBaseParameters;
-	CBaseParameters parameters((_dist_type)(-1), u32(-1), u32(-1));
+	CBaseParameters parameters(_dist_type(u32(-1)), u32(-1), u32(-1));
 	path.clear_not_free();
 	R_ASSERT2(
 		ai().graph_engine().search(
